feat: added sign up functionality with firebase
